add tanstack query provider
