+++ src/App.tsx
@@ -4,7 +4,7 @@ import 'text-encoding-polyfill'
 import 'expo-dev-client'
 
 import { useFonts } from 'expo-font'
-import { Provider } from 'lib/tamagui'
+import { Provider } from 'lib/provider'
 import { FC } from 'react'
 import { LogBox } from 'react-native'
 import PolyfillCrypto from 'react-native-webview-crypto'
