+++ src/lib/provider/tamagui.config.ts
